requirements.txt
python-telegram-bot[job_queue]==20.0
psycopg2-binary
python-dotenv
requests
pillow
